+++ setup.cfg
@@ -70,6 +70,7 @@ colcon_argcomplete.argcomplete_completer =
     catkin_cmake_args = colcon_ros.argcomplete_completer.catkin_cmake_args:CatkinCmakeArgcompleteCompleter
 colcon_core.package_augmentation =
     ros = colcon_ros.package_identification.ros:RosPackageIdentification
+    ros_ament_python = colcon_ros.package_augmentation.ros_ament_python:RosAmentPythonPackageAugmentation
 colcon_core.package_identification =
     ignore_ament_install = colcon_ros.package_identification.ignore:IgnorePackageIdentification
     ros = colcon_ros.package_identification.ros:RosPackageIdentification
